+++ src/card/sets/y2012/dark_ascension.rs
@@ -3373,6 +3373,7 @@ pub(in crate::card::sets) static HELVAULT: CardRecord = CardRecord::new_with_leg
                 })],
                 EffectDef::ExileLinkedToSource {
                     object: EffectRecipientDef::Target(TargetIndex::PRIMARY),
+face_down: false,
 then: None,
 },
             ),
@@ -3390,6 +3391,7 @@ then: None,
                 })],
                 EffectDef::ExileLinkedToSource {
                     object: EffectRecipientDef::Target(TargetIndex::PRIMARY),
+face_down: false,
 then: None,
 },
             ),