+++ src/card/sets/y2019/modern_horizons.rs
@@ -26,6 +26,7 @@ static NOT_YOUR_TURN: TriggerConditionDef =
 static EPHEMERATE_BLINKS: [EffectDef; 3] = [
     EffectDef::ExileLinkedToSource {
         object: EffectRecipientDef::Target(TargetIndex::PRIMARY),
+        face_down: false,
         then: None,
     },
     EffectDef::ReturnLinkedExiles {