+++ src/card/sets/y2020/zendikar_rising.rs
@@ -77,6 +77,7 @@ static SKYCLAVE_APPARITION_ABILITIES: [AbilityDef; 2] = [
         &A_CHEAP_NONLAND_NONTOKEN,
         EffectDef::ExileLinkedToSource {
             object: EffectRecipientDef::Target(TargetIndex::PRIMARY),
+            face_down: false,
             then: None,
         },
     ),