+++ src/card/sets/y2022/kamigawa_neon_dynasty.rs
@@ -155,6 +155,7 @@ static ONE_ARTIFACT_OR_CREATURE: [AbilityTargetDef; 1] = [AbilityTargetDef::exac
 static TOUCH_EXILES_IT: [EffectDef; 2] = [
     EffectDef::ExileLinkedToSource {
         object: EffectRecipientDef::Target(TargetIndex::PRIMARY),
+        face_down: false,
         then: None,
     },
     EffectDef::InstallTrigger(InstalledTriggerDef::once(&TOUCH_RETURNS_IT)),