+++ src/card/sets/y2023/wilds_of_eldraine.rs
@@ -174,6 +174,7 @@ static CAULDRON_GROWS_A_CREATURE: EffectDef = EffectDef::AddCounters {
 static CAULDRON_EXILES_THEN_GROWS: [EffectDef; 2] = [
     EffectDef::ExileLinkedToSource {
         object: EffectRecipientDef::Target(TargetIndex::PRIMARY),
+        face_down: false,
         then: None,
     },
     EffectDef::IfCondition {